(v2): Migrate Caddy to v2 & replace config parser
