+++ dockerv2.go
@@ -76,12 +76,12 @@ func (d *Dockerv2) Redirect() error {
 		if err != nil {
 			return err
 		}
-		d.rw.Header().Add("Cache-Control", fmt.Sprintf("max-age=%d", status301CacheAge))
+		d.rw.Header().Add("Cache-Control", fmt.Sprintf("max-age=%d", Status301CacheAge))
 		d.rw.Header().Add("Status-Code", strconv.Itoa(http.StatusMovedPermanently))
 		http.Redirect(d.rw, d.req, uri, http.StatusMovedPermanently)
 		return nil
 	}
-	d.rw.Header().Add("Cache-Control", fmt.Sprintf("max-age=%d", status301CacheAge))
+	d.rw.Header().Add("Cache-Control", fmt.Sprintf("max-age=%d", Status301CacheAge))
 	d.rw.Header().Add("Status-Code", strconv.Itoa(http.StatusMovedPermanently))
 	http.Redirect(d.rw, d.req, d.rec.To, http.StatusMovedPermanently)
 	return nil
