+++ go.mod
@@ -3,31 +3,9 @@ module go.txtdirect.org/txtdirect
 go 1.13
 
 require (
-	github.com/SchumacherFM/mailout v1.3.0
-	github.com/beorn7/perks v1.0.1 // indirect
-	github.com/caddyserver/caddy v1.0.3
-	github.com/captncraig/caddy-realip v0.0.0-20190710144553-6df827e22ab8
-	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0 // indirect
-	github.com/golang/protobuf v1.3.2 // indirect
-	github.com/gorilla/sessions v1.1.3 // indirect
-	github.com/gorilla/websocket v1.4.1 // indirect
-	github.com/juju/ratelimit v1.0.1 // indirect
-	github.com/klauspost/cpuid v1.2.1 // indirect
-	github.com/kr/pretty v0.1.0 // indirect
-	github.com/mholt/certmagic v0.6.2 // indirect
-	github.com/miekg/caddy-prometheus v0.0.0-20190709133612-1fe4cb19becd
-	github.com/miekg/dns v1.1.16
-	github.com/prometheus/client_golang v1.0.0
-	github.com/prometheus/common v0.4.1
-	github.com/prometheus/procfs v0.0.4 // indirect
-	github.com/quasoft/memstore v0.0.0-20180925164028-84a050167438 // indirect
-	github.com/skip2/go-qrcode v0.0.0-20190110000554-dc11ecdae0a9
-	github.com/steambap/captcha v1.3.0 // indirect
-	golang.org/x/image v0.0.0-20190424155947-59b11bec70c7 // indirect
-	golang.org/x/sys v0.0.0-20190813064441-fde4db37ae7a // indirect
-	golang.org/x/text v0.3.2 // indirect
-	gopkg.in/alexcesaro/quotedprintable.v3 v3.0.0-20150716171945-2caba252f4dc // indirect
-	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
-	gopkg.in/gomail.v2 v2.0.0-20160411212932-81ebce5c23df // indirect
+	github.com/caddyserver/caddy/v2 v2.1.1
+	github.com/miekg/dns v1.1.27
+	github.com/prometheus/common v0.7.0
+	github.com/skip2/go-qrcode v0.0.0-20200617195104-da1b6568686e
 	gopkg.in/natefinch/lumberjack.v2 v2.0.0
 )
