+++ gometa.go
@@ -16,7 +16,6 @@ package txtdirect
 import (
 	"html/template"
 	"net/http"
-	"strconv"
 	"strings"
 )
 
@@ -58,7 +57,7 @@ func (g *Gometa) Serve() error {
 
 	gosource := strings.Contains(g.rec.To, "github.com")
 
-	RequestsByStatus.WithLabelValues(g.req.Host, strconv.Itoa(http.StatusFound)).Add(1)
+	// RequestsByStatus.WithLabelValues(g.req.Host, strconv.Itoa(http.StatusFound)).Add(1)
 	return tmpl.Execute(g.rw, struct {
 		Host        string
 		Path        string
