+++ host.go
@@ -35,12 +35,12 @@ func (h *Host) Redirect() error {
 	}
 	log.Printf("[txtdirect]: %s > %s", h.req.Host+h.req.URL.Path, to)
 	if code == http.StatusMovedPermanently {
-		h.rw.Header().Add("Cache-Control", fmt.Sprintf("max-age=%d", status301CacheAge))
+		h.rw.Header().Add("Cache-Control", fmt.Sprintf("max-age=%d", Status301CacheAge))
 	}
 	h.rw.Header().Add("Status-Code", strconv.Itoa(code))
 	http.Redirect(h.rw, h.req, to, code)
-	if h.c.Prometheus.Enable {
-		RequestsByStatus.WithLabelValues(h.req.Host, strconv.Itoa(code)).Add(1)
-	}
+	// if h.c.Prometheus.Enable {
+	// 	RequestsByStatus.WithLabelValues(h.req.Host, strconv.Itoa(code)).Add(1)
+	// }
 	return nil
 }
